Add day 5
This quickly turned into an unholy mess because of (wrong)
assumptions and a complete lack of thought that went into making the
final day 2 solution. The first part my code was not prepared for was
immediate mode parameters and later it didn't work well with writing to
position mode parameters. Then came the jump statements and I just had
to go back to using an embedded `case` statement because it was easier
than passing extra parameters around.

I might clean this up later but the best solution would probably still
be to just wrap all of this in some class. It works for now and will be
really easy to extend if this interpreter comes up again.